separate createPreviewData function. Rename key organizations
