+++ src/features/import/utils/prepareImportOperations.ts
@@ -3,7 +3,7 @@ import { CellData, ColumnKind, Sheet } from './types';
 export type ZetkinPersonImportOp = {
   data?: Record<string, CellData>;
   op: 'person.import';
-  organizations?: number[];
+  organization?: number[];
   tags?: number[];
 };
 
@@ -65,7 +65,7 @@ export default function prepareImportOperations(
         if (column.kind === ColumnKind.ORGANIZATION) {
           column.mapping.forEach((mappedColumn) => {
             if (mappedColumn.value === row.data[colIdx]) {
-              personImportOps[rowIndex].organizations = [
+              personImportOps[rowIndex].organization = [
                 mappedColumn.orgId as number,
               ];
             }
@@ -75,56 +75,4 @@ export default function prepareImportOperations(
     }
   });
   return personImportOps;
-}
-
-export function prepareImportOperationsForRow(
-  configuredSheet: Sheet,
-  personIndex: number
-) {
-  const personImportOp: Partial<ZetkinPersonImportOp> = {};
-  const row = configuredSheet.rows[personIndex].data;
-
-  configuredSheet.columns.forEach((column, colIdx) => {
-    if (column.selected) {
-      personImportOp.op = 'person.import';
-    }
-    //ID column and fields
-    if (
-      column.kind === ColumnKind.ID_FIELD ||
-      column.kind === ColumnKind.FIELD
-    ) {
-      const fieldKey =
-        column.kind === ColumnKind.ID_FIELD ? column.idField : column.field;
-
-      if (row[colIdx]) {
-        personImportOp.data = {
-          ...personImportOp.data,
-          [`${fieldKey}`]: row[colIdx],
-        };
-      }
-    }
-    //tags
-    if (column.kind === ColumnKind.TAG) {
-      column.mapping.forEach((mappedColumn) => {
-        if (mappedColumn.value === row[colIdx]) {
-          if (!personImportOp.tags) {
-            personImportOp.tags = [];
-          }
-          personImportOp.tags = [
-            ...new Set(personImportOp.tags.concat(mappedColumn.tagIds)),
-          ];
-        }
-      });
-    }
-    //orgs
-    if (column.kind === ColumnKind.ORGANIZATION) {
-      column.mapping.forEach((mappedColumn) => {
-        if (mappedColumn.value === row[colIdx]) {
-          personImportOp.organizations = [mappedColumn?.orgId as number];
-        }
-      });
-    }
-  });
-
-  return personImportOp;
 }
